pyproject.toml
# This is the modern, standardized project metadata section (PEP 621)
[project]
name = "c" # Changed from "t" to match your repo name, as is standard
version = "0.1.0"
description = "A unified toolkit for AI, media, and system utilities."
authors = [
    {name = "TechEliteAutomation", email = "at253341@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"  # <-- CRITICAL FIX APPLIED HERE
dependencies = [
    "google-generativeai>=0.5.0",
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "toml (>=0.10.2,<0.11.0)",
    "textual[dev] (>=3.5.0,<4.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
]

# This section is required by the build system
[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

# --- Poetry Specific Configuration ---

[tool.poetry]
name = "c" # Ensure this matches the project name
version = "0.1.0" # Ensure this matches the project version
description = "A unified toolkit for AI, media, and system utilities."
authors = ["TechEliteAutomation <at253341@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "toolkit", from = "src"}]

[tool.poetry.dependencies]
# The main python version for poetry
python = "^3.13"
# Main dependencies are now read from the [project] table above,
# so this list can be kept in sync or left empty if using a modern Poetry version.
# For compatibility, we list them here as well.
google-generativeai = ">=0.5.0"
python-dotenv = ">=1.0.0"
PyPDF2 = ">=3.0.0"
requests = ">=2.31.0"
beautifulsoup4 = ">=4.12.0"

# This is the ONLY place dev dependencies should be defined
[tool.poetry.group.dev.dependencies]
ruff = ">=0.4.0"
pytest = ">=8.0.0"

# --- Tool Configurations ---

# This is the ONLY tool configuration needed now
[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
# Emulate flake8, isort, and more
select = ["E", "F", "W", "I"]

[tool.ruff.format]
# Ruff's formatter is black-compatible by default
quote-style = "double"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
